// angular/src/services/interfaces/models.ts
export interface CustomerDTO {
    id?: number;
    accountNumber?: number;
    customerName?: string | undefined;
    discount?: number | undefined;
    taxRate?: number | undefined;
    paymentTermId?: number | undefined;
    priceLevelId?: number | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    zone?: string | undefined;
    contactName?: string | undefined;
    phoneNumber?: string | undefined;
    faxNumber?: string | undefined;
    email?: string | undefined;
    isHoldAccount?: boolean;
    creditLimit?: number;
    isBypassCreditLimit?: boolean;
    overBalance?: number;
    statementAmount?: number;
    contacts?: Contact[] | undefined;
    customerCredit?: CustomerCredit;
    locations?: Location[] | undefined;
}
// ...
export interface Location {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    phoneNumber?: string | undefined;
    faxNumber?: string | undefined;
    email?: string | undefined;
    latitude?: string | undefined;
    longitude?: string | undefined;
    customerId?: number;
    customer?: Customer;
    locationTypeId?: number;
    locationType?: LocationType;
    locationName?: string | undefined;
    locationCode?: string | undefined;
    notes?: string | undefined;
    zone?: string | undefined;
}
// ...
export interface Order {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    customerId?: number;
    customer?: Customer;
    paymentTermId?: number | undefined;
    paymentTerm?: PaymentTerm;
    warehouseId?: number;
    warehouse?: Warehouse;
    orderNumber?: number | undefined;
    quoteNumber?: number | undefined;
    customerName?: string | undefined;
    accountNumber?: number;
    paymentTermName?: string | undefined;
    priceLevelId?: number;
    priceLevelName?: string | undefined;
    warehouseName?: string | undefined;
    orderDate?: moment.Moment;
    user?: string | undefined;
    orderStatusId?: number;
    orderStatusName?: string | undefined;
    orderedBy?: string | undefined;
    orderedByEmail?: string | undefined;
    orderedByPhoneNumber?: string | undefined;
    orderedByNotes?: string | undefined;
    deliveryType?: number;
    deliveryDate?: moment.Moment;
    deliveryRoute?: number | undefined;
    discount?: number | undefined;
    restockingFee?: number | undefined;
    restockingAmount?: number | undefined;
    taxRate?: number | undefined;
    totalTax?: number | undefined;
    subTotalAmount?: number | undefined;
    totalAmount?: number | undefined;
    isQuote?: boolean;
    isPrinted?: boolean;
    isCustomerOrder?: boolean;
    linkedInvoiceNumber?: string | undefined;
    orderDetails?: OrderDetail[] | undefined;
    email?: string | undefined;
    phoneNumber?: string | undefined;
    billAddress?: string | undefined;
    billCity?: string | undefined;
    billState?: string | undefined;
    billZipCode?: string | undefined;
    billZone?: string | undefined;
    billPhoneNumber?: string | undefined;
    billContactName?: string | undefined;
    shipAddressName?: string | undefined;
    shipAddress?: string | undefined;
    shipCity?: string | undefined;
    shipState?: string | undefined;
    shipZipCode?: string | undefined;
    shipZone?: string | undefined;
    shipPhoneNumber?: string | undefined;
    shipContactName?: string | undefined;
    isHoldInvoice?: boolean;
    invoiceNumber?: number | undefined;
    fwd?: number | undefined;
    originalInvoiceNumber?: number | undefined;
    orderType?: number | undefined;
    handlingType?: number | undefined;
    creditType?: number | undefined;
    singleInvoice?: number | undefined;
    purchaseOrderNumber?: string | undefined;
    type?: string | undefined;
    category?: string | undefined;
    carrier?: string | undefined;
    trackingNumber?: string | undefined;
    salesRepresentative1?: string | undefined;
    salesRepresentative2?: string | undefined;
    invoiceNotes?: string | undefined;
    particular?: string | undefined;
    paymentReference?: string | undefined;
    postedBy?: string | undefined;
    accountingPaidBy?: string | undefined;
    transactionDate?: moment.Moment | undefined;
    postDate?: moment.Moment | undefined;
    paymentDate?: moment.Moment | undefined;
    debit?: number | undefined;
    credit?: number | undefined;
    quantity?: number | undefined;
    quantityShipped?: number | undefined;
    currentCost?: number | undefined;
    amountPaid?: number | undefined;
    balance?: number | undefined;
    payment?: number | undefined;
    listPrice?: number | undefined;
    rgaType?: number;
    rgaReason?: number;
    rgaReasonNotes?: string | undefined;
}

export interface OrderDetail {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    orderId?: number;
    order?: Order;
    productId?: number;
    product?: Product;
    categoryId?: number;
    orderQuantity?: number;
    onHandQuantity?: number | undefined;
    location?: string | undefined;
    warehouseLocationId?: number | undefined;
    partNumber?: string | undefined;
    partDescription?: string | undefined;
    brand?: string | undefined;
    mainPartsLinkNumber?: string | undefined;
    mainOEMNumber?: string | undefined;
    vendorCode?: string | undefined;
    vendorPartNumber?: string | undefined;
    vendorPrice?: number;
    vendorOnHand?: number;
    yearFrom?: number;
    yearTo?: number;
    listPrice?: number;
    wholesalePrice?: number;
    discountedPrice?: number;
    totalAmount?: number;
    restockingFee?: number | undefined;
    restockingAmount?: number | undefined;
    discount?: number;
    discountAmount?: number;
    discountRecord?: number;
    discountRecordAmount?: number;
    partsLinks?: string | undefined;
    oeMs?: string | undefined;
    vendorCodes?: string | undefined;
    imageUrl?: string | undefined;
    stocks?: WarehouseStock[] | undefined;
    vendorCatalogs?: VendorCatalog[] | undefined;
    warehouseTrackings?: WarehouseTracking[] | undefined;
    vehicle?: string | undefined;
    salesRepresentative?: string | undefined;
    warehouseLocation?: string | undefined;
    vendorInfo?: string | undefined;
    warehouseTracking?: string | undefined;
    carrier?: string | undefined;
    trackingNumber?: string | undefined;
    salesOrderNumber?: number | undefined;
    originalInvoiceNumber?: number | undefined;
    buyOutOrder?: number | undefined;
    partSize?: number | undefined;
    statusId?: number | undefined;
    shippedQuantity?: number | undefined;
    scannedQuantity?: number | undefined;
    pickedQuantity?: number | undefined;
    returnQuantity?: number | undefined;
    remainingQuantity?: number | undefined;
    itemCatalog?: number | undefined;
    toShip?: number | undefined;
    price?: number | undefined;
    unitCost?: number | undefined;
    shipCharge?: number | undefined;
    shipDate?: moment.Moment | undefined;
    rgaInspectedCode?: number | undefined;
    rgaState?: number | undefined;
    rgaLocation?: string | undefined;
    rgaPartNumber?: string | undefined;
    isCreditMemoCreated?: boolean;
}
// ...
export interface PaymentHistoryDTO {
    id?: number;
    paymentDate?: moment.Moment;
    totalAmountDue?: number;
    paymentAmount?: number;
    paymentBalance?: number;
    customerCreditAmountUsed?: number;
    paymentType?: string | undefined;
    referenceNumber?: string | undefined;
    invoiceAmount?: number;
    invoicePaymentAmount?: number;
    invoiceBalance?: number;
    linkedInvoiceNumber?: string | undefined;
}

export interface PaymentPaginatedListDTO {
    recordCount?: number;
    data?: Payment[] | undefined;
}

export interface PaymentTerm {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    termName?: string | undefined;
    description?: string | undefined;
    numberOfDays?: number;
    notes?: string | undefined;
}
// ...
export interface ProductDTO {
    id?: number;
    partNumber?: string | undefined;
    partDescription?: string | undefined;
    brand?: string | undefined;
    categoryId?: number | undefined;
    sequenceId?: number | undefined;
    statusId?: number | undefined;
    partSizeId?: number | undefined;
    partsLinks?: string | undefined;
    oeMs?: string | undefined;
    vendorCodes?: string | undefined;
    partsLinkNumber?: string | undefined;
    oemNumber?: string | undefined;
    yearFrom?: number;
    yearTo?: number;
    listPrice?: number | undefined;
    priceLevel1?: number | undefined;
    priceLevel2?: number | undefined;
    priceLevel3?: number | undefined;
    priceLevel4?: number | undefined;
    priceLevel5?: number | undefined;
    priceLevel6?: number | undefined;
    priceLevel7?: number | undefined;
    priceLevel8?: number | undefined;
    stock?: number;
    currentCost?: number;
    stocks?: WarehouseStock[] | undefined;
    vendorCatalogs?: VendorCatalog[] | undefined;
}

export interface ProductDTOPaginatedListDTO {
    recordCount?: number;
    data?: ProductDTO[] | undefined;
}

export interface ProductFilterDTO {
    state?: number;
    year?: number | undefined;
    categoryIds?: number[] | undefined;
    sequenceIds?: number[] | undefined;
    make?: string | undefined;
    model?: string | undefined;
    isCAProduct?: boolean;
    isNVProduct?: boolean;
}
// ...
export interface Role {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    name?: string | undefined;
    code?: string | undefined;
    description?: string | undefined;
    sortOrder?: number;
    userTypeId?: number;
    userType?: UserType;
    notes?: string | undefined;
    rolePermissions?: RolePermission[] | undefined;
}
// ...
export interface User {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    roleId?: number;
    role?: Role;
    customerId?: number | undefined;
    customer?: Customer;
    locationId?: number | undefined;
    location?: Location;
    userName?: string | undefined;
    email?: string | undefined;
    isCustomerUser?: boolean;
    isActivated?: boolean;
    isChangePasswordOnLogin?: boolean;
    passwordHash?: string | undefined;
    passwordSalt?: string | undefined;
}
// ...
export interface VendorCatalog {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    vendorCode?: string | undefined;
    vendorPartNumber?: string | undefined;
    partsLinkNumber?: string | undefined;
    price?: number;
    onHand?: number;
    cutoffTime?: moment.Moment;
}
// ...
export interface Warehouse {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    warehouseName?: string | undefined;
    description?: string | undefined;
    notes?: string | undefined;
}